+++ src/lib/ng-qubee.module.ts
@@ -4,10 +4,12 @@ import { DriverEnum } from './enums/driver.enum';
 import { IConfig } from './interfaces/config.interface';
 import { IRequestStrategy } from './interfaces/request-strategy.interface';
 import { IResponseStrategy } from './interfaces/response-strategy.interface';
-import { NestjsResponseOptions } from './models/response-options';
+import { JsonApiResponseOptions, NestjsResponseOptions } from './models/response-options';
 import { NgQubeeService } from './services/ng-qubee.service';
 import { NestService } from './services/nest.service';
 import { PaginationService } from './services/pagination.service';
+import { JsonApiRequestStrategy } from './strategies/json-api-request.strategy';
+import { JsonApiResponseStrategy } from './strategies/json-api-response.strategy';
 import { LaravelRequestStrategy } from './strategies/laravel-request.strategy';
 import { LaravelResponseStrategy } from './strategies/laravel-response.strategy';
 import { NestjsRequestStrategy } from './strategies/nestjs-request.strategy';
@@ -23,6 +25,8 @@ import { SpatieResponseStrategy } from './strategies/spatie-response.strategy';
  */
 function resolveRequestStrategy(driver: DriverEnum): IRequestStrategy {
   switch (driver) {
+    case DriverEnum.JSON_API:
+      return new JsonApiRequestStrategy();
     case DriverEnum.NESTJS:
       return new NestjsRequestStrategy();
     case DriverEnum.SPATIE:
@@ -40,6 +44,8 @@ function resolveRequestStrategy(driver: DriverEnum): IRequestStrategy {
  */
 function resolveResponseStrategy(driver: DriverEnum): IResponseStrategy {
   switch (driver) {
+    case DriverEnum.JSON_API:
+      return new JsonApiResponseStrategy();
     case DriverEnum.NESTJS:
       return new NestjsResponseStrategy();
     case DriverEnum.SPATIE:
@@ -84,6 +90,10 @@ export class NgQubeeModule {
           provide: PaginationService,
           useFactory: () => {
             const responseConfig = Object.assign({}, config.response);
+
+            if (driver === DriverEnum.JSON_API) {
+              return new PaginationService(responseStrategy, new JsonApiResponseOptions(responseConfig));
+            }
 
             return driver === DriverEnum.NESTJS
               ? new PaginationService(responseStrategy, new NestjsResponseOptions(responseConfig))