+++ src/lib/services/pagination.service.ts
@@ -1,12 +1,9 @@
-import { Inject, Injectable, Optional } from '@angular/core';
-
 import { IPaginatedObject } from '../interfaces/paginated-object.interface';
 import { IPaginationConfig } from '../interfaces/pagination-config.interface';
 import { IResponseStrategy } from '../interfaces/response-strategy.interface';
 import { PaginatedCollection } from '../models/paginated-collection';
 import { ResponseOptions } from '../models/response-options';
 
-@Injectable()
 export class PaginationService {
 
   /**
@@ -20,8 +17,8 @@ export class PaginationService {
   private _responseStrategy: IResponseStrategy;
 
   constructor(
-    @Inject('RESPONSE_STRATEGY') responseStrategy: IResponseStrategy,
-    @Inject('RESPONSE_OPTIONS') @Optional() options: IPaginationConfig = {}
+    responseStrategy: IResponseStrategy,
+    options: IPaginationConfig = {}
   ) {
     this._options = new ResponseOptions(options);
     this._responseStrategy = responseStrategy;